Decouple Orleans server dependency from base CloudActors
We already had a separate assembly for the Orleans server extensions. By moving the grain generator to its own assembly, we can fully decouple the server/hosting interface and types from the bare business logic implemented via cloud actors. The only dependency would be the serializer generation which needs to be based on annotating the actual types, so it's unavoidable. But this is just the abstractions package.
