+++ src/CloudActors.CodeAnalysis/ActorState.sbntxt
