+++ src/CloudActors.CodeAnalysis/EventSourced.sbntxt
