+++ src/CloudActors.CodeAnalysis/Extensions.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace Devlooped.CloudActors;
+
+static class Extensions
+{
+}
