+++ src/CloudActors.Orleans.CodeAnalysis/ActorGrain.sbntxt
