+++ src/CloudActors.Package/CloudActors.Package.msbuildproj
@@ -9,9 +9,8 @@
     <PackageReference Include="NuGetizer" Version="1.2.2" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\CloudActors.CodeAnaysis\CloudActors.CodeAnaysis.csproj" />
+    <ProjectReference Include="..\CloudActors.CodeAnalysis\CloudActors.CodeAnalysis.csproj" />
     <ProjectReference Include="..\CloudActors.CodeFix\CloudActors.CodeFix.csproj" />
     <ProjectReference Include="..\CloudActors\CloudActors.csproj" />
-    <ProjectReference Include="..\CloudActors.Orleans\CloudActors.Orleans.csproj" />
   </ItemGroup>
 </Project>
