+++ src/Tests/Tests.csproj
@@ -31,8 +31,9 @@
     <ProjectReference Include="..\CloudActors.Orleans\CloudActors.Orleans.csproj" />
     <ProjectReference Include="..\CloudActors.Streamstone\CloudActors.Streamstone.csproj" />
     <ProjectReference Include="..\CloudActors\CloudActors.csproj" />
-    <ProjectReference Include="..\CloudActors.CodeAnaysis\CloudActors.CodeAnaysis.csproj" OutputItemType="Analyzer" />
+    <ProjectReference Include="..\CloudActors.CodeAnalysis\CloudActors.CodeAnalysis.csproj" OutputItemType="Analyzer" />
     <ProjectReference Include="..\CloudActors.CodeFix\CloudActors.CodeFix.csproj" OutputItemType="Analyzer" />
+    <ProjectReference Include="..\CloudActors.Orleans.CodeAnalysis\CloudActors.Orleans.CodeAnalysis.csproj" OutputItemType="Analyzer" />
     <ProjectReference Include="..\TestDomain\TestDomain.csproj" />
   </ItemGroup>
